+++ src/native/include/xaudio2fx.h
@@ -1,4 +1,4 @@
 #ifdef _MSC_VER
 #pragma once
 #endif
-#include "xaudio2Redist.h"
+#include "xaudio2fxRedist.h"